+++ poem-openapi-derive/src/oneof.rs
@@ -97,6 +97,8 @@ pub(crate) fn generate(args: DeriveInput) -> GeneratorResult<TokenStream> {
 
             type RawValueType = Self;
 
+            type RawElementValueType = Self;
+
             fn name() -> ::std::borrow::Cow<'static, str> {
                 ::std::convert::Into::into("object")
             }
@@ -122,6 +124,10 @@ pub(crate) fn generate(args: DeriveInput) -> GeneratorResult<TokenStream> {
 
             fn as_raw_value(&self) -> ::std::option::Option<&Self::RawValueType> {
                 ::std::option::Option::Some(self)
+            }
+
+            fn raw_element_iter<'a>(&'a self) -> ::std::boxed::Box<dyn ::std::iter::Iterator<Item = &'a Self::RawElementValueType> + 'a> {
+                ::std::boxed::Box::new(::std::iter::IntoIterator::into_iter(self.as_raw_value()))
             }
         }