+++ poem-openapi-derive/src/validators.rs
@@ -1,5 +1,5 @@
 use darling::{util::SpannedValue, FromMeta};
-use proc_macro2::TokenStream;
+use proc_macro2::{Span, TokenStream};
 use quote::quote;
 use regex::Regex;
 use syn::{Error, Type};
@@ -11,6 +11,7 @@ use crate::{
 
 #[derive(FromMeta, Default, Clone)]
 pub(crate) struct Validators {
+    // for elements
     #[darling(default)]
     multiple_of: Option<SpannedValue<f64>>,
     #[darling(default)]
@@ -23,14 +24,18 @@ pub(crate) struct Validators {
     min_length: Option<SpannedValue<usize>>,
     #[darling(default)]
     pattern: Option<SpannedValue<String>>,
+
+    // for containers
     #[darling(default)]
     max_items: Option<SpannedValue<usize>>,
     #[darling(default)]
     min_items: Option<SpannedValue<usize>>,
     #[darling(default)]
-    unique_items: bool,
+    unique_items: Option<SpannedValue<bool>>,
     #[darling(default)]
-    list: SpannedValue<bool>,
+    max_properties: Option<SpannedValue<usize>>,
+    #[darling(default)]
+    min_properties: Option<SpannedValue<usize>>,
 }
 
 impl Validators {
@@ -38,8 +43,12 @@ impl Validators {
         &self,
         crate_name: &TokenStream,
     ) -> GeneratorResult<(Vec<TokenStream>, Vec<TokenStream>)> {
-        let mut list_validators = Vec::new();
+        let mut container_validators = Vec::new();
         let mut elem_validators = Vec::new();
+
+        //////////////////////////////////////////////////////////////////////////////
+        // element validators
+        //////////////////////////////////////////////////////////////////////////////
 
         if let Some(value) = self.multiple_of {
             // https://datatracker.ietf.org/doc/html/draft-wright-json-schema-validation-00#section-5.1
@@ -89,68 +98,78 @@ impl Validators {
             elem_validators.push(quote!(#crate_name::validation::Pattern::new(#value)));
         }
 
+        //////////////////////////////////////////////////////////////////////////////
+        // container validators
+        //////////////////////////////////////////////////////////////////////////////
+
         if let Some(value) = self.max_items {
             // https://datatracker.ietf.org/doc/html/draft-wright-json-schema-validation-00#section-5.10
             let value = &*value;
-            list_validators.push(quote!(#crate_name::validation::MaxItems::new(#value)));
+            container_validators.push(quote!(#crate_name::validation::MaxItems::new(#value)));
         }
 
         if let Some(value) = self.min_items {
             // https://datatracker.ietf.org/doc/html/draft-wright-json-schema-validation-00#section-5.11
             let value = &*value;
-            list_validators.push(quote!(#crate_name::validation::MinItems::new(#value)));
+            container_validators.push(quote!(#crate_name::validation::MinItems::new(#value)));
         }
 
-        if self.unique_items {
+        if self.unique_items.map(|value| *value).unwrap_or_default() {
             // https://datatracker.ietf.org/doc/html/draft-wright-json-schema-validation-00#section-5.12
-            list_validators.push(quote!(#crate_name::validation::UniqueItems::new()));
+            container_validators.push(quote!(#crate_name::validation::UniqueItems::new()));
         }
 
-        Ok((list_validators, elem_validators))
+        if let Some(value) = self.max_properties {
+            // https://json-schema.org/draft/2020-12/json-schema-validation.html#rfc.section.6.5.1
+            let value = &*value;
+            container_validators.push(quote!(#crate_name::validation::MaxProperties::new(#value)));
+        }
+
+        if let Some(value) = self.min_properties {
+            // https://json-schema.org/draft/2020-12/json-schema-validation.html#rfc.section.6.5.2
+            let value = &*value;
+            container_validators.push(quote!(#crate_name::validation::MinProperties::new(#value)));
+        }
+
+        Ok((container_validators, elem_validators))
+    }
+
+    fn first_container_validator_span(&self) -> Option<Span> {
+        self.max_items
+            .as_ref()
+            .map(SpannedValue::span)
+            .or(self.min_items.as_ref().map(SpannedValue::span))
+            .or(self.unique_items.as_ref().map(SpannedValue::span))
+            .or(self.max_properties.as_ref().map(SpannedValue::span))
+            .or(self.min_properties.as_ref().map(SpannedValue::span))
     }
 
     pub(crate) fn create_obj_field_checker(
         &self,
         crate_name: &TokenStream,
         field_name: &str,
-    ) -> GeneratorResult<Option<TokenStream>> {
-        let (list_validators, elem_validators) = self.create_validators(crate_name)?;
-        if !*self.list {
-            let validators = list_validators.into_iter().chain(elem_validators);
-            Ok(Some(quote! {
-                #(
-                    if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {
-                        let validator = #validators;
-                        if !#crate_name::validation::Validator::check(&validator, value) {
-                            return Err(#crate_name::types::ParseError::<Self>::custom(format!("field `{}` verification failed. {}", #field_name, validator)));
-                        }
-                    }
-                )*
-            }))
-        } else {
-            Ok(Some(quote! {
-                #(
-                    if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {
-                        let validator = #list_validators;
-                        if !#crate_name::validation::Validator::check(&validator, value) {
-                            return Err(#crate_name::types::ParseError::<Self>::custom(format!("field `{}` verification failed. {}", #field_name, validator)));
-                        }
-                    }
-                )*
-                #(
-                    if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {
-                        for item in value {
-                            if let ::std::option::Option::Some(item) = #crate_name::types::Type::as_raw_value(item) {
-                                let validator = #elem_validators;
-                                if !#crate_name::validation::Validator::check(&validator, item) {
-                                    return Err(#crate_name::types::ParseError::<Self>::custom(format!("field `{}` verification failed. {}", #field_name, validator)));
-                                }
-                            }
-                        }
-                    }
-                )*
-            }))
-        }
+    ) -> GeneratorResult<TokenStream> {
+        let (container_validators, elem_validators) = self.create_validators(crate_name)?;
+
+        Ok(quote! {
+            #(
+            for elem in #crate_name::types::Type::raw_element_iter(&value) {
+                let validator = #elem_validators;
+                if !#crate_name::validation::Validator::check(&validator, elem) {
+                    return Err(#crate_name::types::ParseError::<Self>::custom(format!("field `{}` verification failed. {}", #field_name, validator)));
+                }
+            }
+            )*
+
+            #(
+            if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {         let
+                validator = #container_validators;
+                if !#crate_name::validation::Validator::check(&validator, value) {
+                    return Err(#crate_name::types::ParseError::<Self>::custom(format!("field `{}` verification failed. {}", #field_name, validator)));
+                }
+            }
+            )*
+        })
     }
 
     pub(crate) fn create_param_checker(
@@ -159,33 +178,38 @@ impl Validators {
         res_ty: &Type,
         arg_name: &str,
     ) -> GeneratorResult<Option<TokenStream>> {
-        let (list_validators, elem_validators) = self.create_validators(crate_name)?;
-        if *self.list {
+        let (container_validators, elem_validators) = self.create_validators(crate_name)?;
+
+        if !container_validators.is_empty() {
             return Err(Error::new(
-                self.list.span(),
-                "The `list` attribute is not supported for parameters.",
+                self.first_container_validator_span().unwrap(),
+                "The `container` validators is not supported for parameters.",
             )
             .into());
         }
 
-        let validators = list_validators.into_iter().chain(elem_validators);
+        if elem_validators.is_empty() {
+            return Ok(None);
+        }
+
         Ok(Some(quote! {
             #(
-                if let ::std::option::Option::Some(item) = #crate_name::types::Type::as_raw_value(value) {
-                    let validator = #validators;
-                    if !#crate_name::validation::Validator::check(&validator, value) {
-                        let err = #crate_name::ParseRequestError::ParseParam {
-                            name: #arg_name,
-                            reason: ::std::format!("verification failed. {}", validator),
-                        };
-                        if <#res_ty as #crate_name::ApiResponse>::BAD_REQUEST_HANDLER {
-                            let resp = <#res_ty as #crate_name::ApiResponse>::from_parse_request_error(err);
-                            return #crate_name::__private::poem::IntoResponse::into_response(resp);
-                        } else {
-                            return #crate_name::__private::poem::IntoResponse::into_response(err);
-                        }
+            if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {
+                let validator = #elem_validators;
+                if !#crate_name::validation::Validator::check(&validator, value) {
+                    let err = #crate_name::ParseRequestError::ParseParam {
+                        name: #arg_name,
+                        reason: ::std::format!("verification failed. {}", validator),
+                    };
+
+                    if <#res_ty as #crate_name::ApiResponse>::BAD_REQUEST_HANDLER {
+                        let resp = <#res_ty as #crate_name::ApiResponse>::from_parse_request_error(err);
+                        return #crate_name::__private::poem::IntoResponse::into_response(resp);
+                    } else {
+                        return #crate_name::__private::poem::IntoResponse::into_response(err);
                     }
                 }
+            }
             )*
         }))
     }
@@ -194,81 +218,74 @@ impl Validators {
         &self,
         crate_name: &TokenStream,
         field_name: &str,
-    ) -> GeneratorResult<Option<TokenStream>> {
-        let (list_validators, elem_validators) = self.create_validators(crate_name)?;
-        if !*self.list {
-            let validators = list_validators.into_iter().chain(elem_validators);
-            Ok(Some(quote! {
-                #(
-                    if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {
-                        let validator = #validators;
-                        if !#crate_name::validation::Validator::check(&validator, value) {
-                            return Err(#crate_name::ParseRequestError::ParseRequestBody(
-                                #crate_name::__private::poem::Response::builder()
-                                    .status(#crate_name::__private::poem::http::StatusCode::BAD_REQUEST)
-                                    .body(::std::format!("field `{}` verification failed. {}", #field_name, validator))
-                            ));
-                        }
+    ) -> GeneratorResult<TokenStream> {
+        let (container_validators, elem_validators) = self.create_validators(crate_name)?;
+
+        Ok(quote! {
+            #(
+            for item in #crate_name::types::Type::raw_element_iter(&value) {
+                let validator = #elem_validators;
+                if !#crate_name::validation::Validator::check(&validator, item) {
+                    return Err(#crate_name::ParseRequestError::ParseRequestBody(
+                        #crate_name::__private::poem::Response::builder()
+                            .status(#crate_name::__private::poem::http::StatusCode::BAD_REQUEST)
+                            .body(::std::format!("field `{}` verification failed. {}", #field_name, validator))
+                        ));
+                }
+            }
+            )*
+
+            #(
+            if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {
+                let validator = #container_validators;
+                if !#crate_name::validation::Validator::check(&validator, value) {
+                    return Err(#crate_name::ParseRequestError::ParseRequestBody(
+                        #crate_name::__private::poem::Response::builder()
+                            .status(#crate_name::__private::poem::http::StatusCode::BAD_REQUEST)
+                            .body(::std::format!("field `{}` verification failed. {}", #field_name, validator))
+                        ));
                     }
-                )*
-            }))
-        } else {
-            Ok(Some(quote! {
-                #(
-                    if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {
-                        let validator = #list_validators;
-                        if !#crate_name::validation::Validator::check(&validator, value) {
-                            return Err(#crate_name::ParseRequestError::ParseRequestBody(
-                                #crate_name::__private::poem::Response::builder()
-                                    .status(#crate_name::__private::poem::http::StatusCode::BAD_REQUEST)
-                                    .body(::std::format!("field `{}` verification failed. {}", #field_name, validator))
-                            ));
-                        }
-                    }
-                )*
-                #(
-                    if let ::std::option::Option::Some(value) = #crate_name::types::Type::as_raw_value(&value) {
-                        let validator = #elem_validators;
-                        for item in value {
-                            if let ::std::option::Option::Some(item) = #crate_name::types::Type::as_raw_value(item) {
-                                if !#crate_name::validation::Validator::check(&validator, item) {
-                                    return Err(#crate_name::ParseRequestError::ParseRequestBody(
-                                        #crate_name::__private::poem::Response::builder()
-                                            .status(#crate_name::__private::poem::http::StatusCode::BAD_REQUEST)
-                                            .body(::std::format!("field `{}` verification failed. {}", #field_name, validator))
-                                    ));
-                                }
-                            }
-                        }
-                    }
-                )*
-            }))
-        }
+                }
+            )*
+        })
     }
 
     pub(crate) fn create_update_meta(
         &self,
         crate_name: &TokenStream,
-    ) -> GeneratorResult<Option<TokenStream>> {
-        let (list_validators, elem_validators) = self.create_validators(crate_name)?;
-        if !*self.list {
-            let validators = list_validators.into_iter().chain(elem_validators);
-            Ok(Some(quote! {
+    ) -> GeneratorResult<TokenStream> {
+        let (container_validators, elem_validators) = self.create_validators(crate_name)?;
+
+        let update_elem_meta = quote! {
+            if original_schema.is_array() {
+                let mut items_schema = #crate_name::registry::MetaSchema::ANY;
                 #(
-                    #crate_name::validation::ValidatorMeta::update_meta(&#validators, &mut schema);
+                #crate_name::validation::ValidatorMeta::update_meta(&#elem_validators, &mut items_schema);
                 )*
-            }))
+                schema.items = ::std::option::Option::Some(::std::boxed::Box::new(#crate_name::registry::MetaSchemaRef::Inline(::std::boxed::Box::new(items_schema))));
+            } else if original_schema.is_object() {
+                let mut additional_properties_schema = #crate_name::registry::MetaSchema::ANY;
+                #(
+                #crate_name::validation::ValidatorMeta::update_meta(&#elem_validators, &mut additional_properties_schema);
+                )*
+                schema.additional_properties = ::std::option::Option::Some(::std::boxed::Box::new(#crate_name::registry::MetaSchemaRef::Inline(::std::boxed::Box::new(additional_properties_schema))));
+            } else {
+                #(
+                #crate_name::validation::ValidatorMeta::update_meta(&#elem_validators, &mut schema);
+                )*
+            }
+        };
+        let update_elem_meta = if !elem_validators.is_empty() {
+            Some(update_elem_meta)
         } else {
-            Ok(Some(quote! {
-                #(
-                    #crate_name::validation::ValidatorMeta::update_meta(&#list_validators, &mut schema);
-                )*
-                #(
-                    let mut items_schema = #crate_name::registry::MetaSchema::ANY;
-                    #crate_name::validation::ValidatorMeta::update_meta(&#elem_validators, &mut items_schema);
-                    schema.items = ::std::option::Option::Some(::std::boxed::Box::new(#crate_name::registry::MetaSchemaRef::Inline(::std::boxed::Box::new(items_schema))));
-                )*
-            }))
-        }
+            None
+        };
+
+        Ok(quote! {
+            #(
+            #crate_name::validation::ValidatorMeta::update_meta(&#container_validators, &mut schema);
+            )*
+            #update_elem_meta
+        })
     }
 }