+++ poem-openapi/src/types/binary.rs
@@ -12,14 +12,28 @@ use crate::{
 pub struct Binary(pub Vec<u8>);
 
 impl Type for Binary {
+    const IS_REQUIRED: bool = true;
+
+    type RawValueType = Self;
+
+    type RawElementValueType = Self;
+
     fn name() -> Cow<'static, str> {
         "string(binary)".into()
     }
 
-    impl_raw_value_type!();
-
     fn schema_ref() -> MetaSchemaRef {
         MetaSchemaRef::Inline(Box::new(MetaSchema::new_with_format("string", "binary")))
+    }
+
+    fn as_raw_value(&self) -> Option<&Self::RawValueType> {
+        Some(self)
+    }
+
+    fn raw_element_iter<'a>(
+        &'a self,
+    ) -> Box<dyn Iterator<Item = &'a Self::RawElementValueType> + 'a> {
+        Box::new(self.as_raw_value().into_iter())
     }
 }