+++ poem-openapi/src/types/macros.rs
@@ -1,8 +1,0 @@
-macro_rules! impl_raw_value_type {
-    () => {
-        type RawValueType = Self;
-        fn as_raw_value(&self) -> Option<&Self::RawValueType> {
-            Some(self)
-        }
-    };
-}