+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Remove the `list` attribute of the validator, it is no longer needed.
+- Add `maxProperties` and `minProperties` validators.
+
 # [1.0.43] 2021-12-07
 
 - Change the schema type of enum to `string` [#121](https://github.com/poem-web/poem/issues/121)